complete fix

+++ src/tools/ftAreaAverage2f.cpp
@@ -3,7 +3,7 @@
 
 namespace flowTools {
 	
-	void ftAreaAverage2f::setup(int _scaleFactor, string _name) {
+	void ftAreaAverage2f::setup(float _scaleFactor, string _name) {
 		scaleFactor = _scaleFactor;
 		
 		quad.getVertices().resize(4);
